Fixed the following bugs: bigmem and smallmem should be read in as longs rather than merely ints workerthread malloc left out allocation for locList when initially generating positions, we could exceed smallmem positions because maxentriesperbuf was calculated wrong (still is but we're not using it)

+++ js2.h
@@ -114,8 +114,8 @@ extern int useLabels;
 extern int directed;
 extern int cutoffStep;
 extern int checkFullTree;
-extern int bigmem;
-extern int smallmem;
+extern long int bigmem;
+extern long int smallmem;
 extern int numThreads;
 extern unsigned int maxEntriesPerBuf;
